backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Pillow==10.1.0
PyJWT==2.8.0
werkzeug==3.0.1
